requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastapi==0.75.0
fastjsonschema==2.15.3
Flask==2.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==7.33.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
pymssql==2.2.4
PyMySQL==1.0.2
pyodbc==4.0.32
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.0.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.32
starlette==0.17.1
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.2.1.post0
typing_extensions==4.1.1
uvicorn==0.17.5
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.7.0
